requirements.txt
Django==4.0.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework
djangorestframework-simplejwt==5.1.0